Move federationapi.readers package into routing
This seems to have been missed when we folded all the readers and
writers packages into routing.
