+++ src/github.com/matrix-org/dendrite/federationapi/routing/version.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package readers
+package routing
 
 import (
 	"github.com/matrix-org/util"
